remove unnecessary duplication of state interpolation
