+++ napari_animation/_tests/test_interpolation.py
@@ -1,4 +1,3 @@
-import numbers
 from dataclasses import asdict
 
 import numpy as np
@@ -9,38 +8,11 @@ from napari_animation.interpolation import (
     interpolate_log,
     interpolate_num,
     interpolate_seq,
-    interpolate_state,
 )
 
-from ..utils import keys_to_list, nested_get
-
+from ..utils import nested_assert_close
 
 # Define some functions used for testing
-def nested_assert_close(a, b):
-    """ Assert close on nested dicts."""
-    a_keys = [key for key in keys_to_list(a)]
-    b_keys = [key for key in keys_to_list(b)]
-
-    assert a_keys == b_keys
-
-    for key in a_keys:
-        a_1 = nested_get(a, key)
-        b_1 = nested_get(b, key)
-
-        nested_seq_assert_close(a_1, b_1)
-
-
-def nested_seq_assert_close(a, b):
-    """ Assert close to scalar or potentially nested qequences of numeric types and others."""
-    if isinstance(a, (list, tuple)) and isinstance(b, (list, tuple)):
-        assert type(a) == type(b)
-        for a_v, b_v in zip(a, b):
-            nested_seq_assert_close(a_v, b_v)
-    else:
-        if isinstance(a, numbers.Number):
-            np.testing.assert_allclose(a, b)
-        else:
-            assert a == b
 
 
 # Actual tests
@@ -87,18 +59,20 @@ def test_interpolate_log(a, b, fraction):
 
 
 @pytest.mark.parametrize("fraction", [0, 0.2, 0.4, 0.6, 0.8, 1])
-def test_interpolate_state(key_frames, fraction):
+def test_interpolate_state(frame_sequence, fraction):
     """Check that state interpolation works"""
-    initial_state = asdict(key_frames[0].viewer_state)
-    final_state = asdict(key_frames[1].viewer_state)
-    result = interpolate_state(initial_state, final_state, fraction)
-    assert len(result) == len(initial_state)
+    initial_state = frame_sequence[0]
+    final_state = frame_sequence[-1]
+    result = frame_sequence._interpolate_state(
+        initial_state, final_state, fraction
+    )
+    assert len(asdict(result)) == len(asdict(initial_state))
     if fraction == 0:
         # assert result == initial_state
-        nested_assert_close(result, initial_state)
+        nested_assert_close(asdict(result), asdict(initial_state))
     elif fraction == 1:
         # assert result == final_state
-        nested_assert_close(result, final_state)
+        nested_assert_close(asdict(result), asdict(final_state))
 
     # else:
     # should find something else to test
